backend/scripts/Selenium_cuisine_du_monde/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-filter==25.2
django-cors-headers==4.9.0
drf-yasg==1.21.14
gunicorn==23.0.0
whitenoise==6.8.2
